installer: port the track build, install and uninstall to C#
Step 3 of tracks/INSTALLER_HANDOFF.md. The C# now derives, registers and
removes Flat Pad on its own; only the GUI and the game unpacker are left.

Validated by comparing the FILES, not the console output: install with the
Python, hash the whole tree, install with the C#, hash again. All 1530 hashes
match — the 1523 track files, both system/*.table registries and the five UI
tiles — from a warm install and from a cold start with no .orig snapshots.
Uninstall matches too, leaving both registries on identical stock hashes. The
C#-built track then passes the Python verifier unchanged.

That comparison earned its keep immediately. Python floats are doubles and
PAD_Y = 18.6 is not representable in binary, so declaring the geometry
constants as C# `float` put every spawn Y one ULP out — identical console
output, different bytes on disk, and nothing short of a hash would have said
so. All geometry constants are now doubles, narrowing to float32 only at the
moment of packing.

Also fixes a real output difference: Python's print(f"\n...") emits CRLF for
the embedded newline on Windows, C#'s Console.WriteLine does not. The log sink
owns the line ending now.

--pr-runoff-spawn is deliberately not ported. It edits a base-game track, which
is the one thing this tool exists to avoid; uninstall still reverts an older
run's snapshot.

51 tests. The new ones guard the control-point trap permanently — reshaping
asserts the point COUNT and that every [3] scale is still (1,1,1), which is
exactly how the naive "offset every vec3" version was caught the first time.

+++ FlatPad.Core/FlatPad/Verifier.cs
@@ -231,7 +231,7 @@ public sealed class Verifier(string gameRoot)
             if (n.Number != 2 || n.Message is null)
                 continue;
             string name = SceneNodes.Name(n);
-            float? want = OvalRadius(name);
+            double? want = OvalRadius(name);
             if (want is null)
                 continue;
